src/requirements.txt
flask==3.1.0
flask-cors==5.0.0
mysql-connector-python==9.3.0
python-dotenv==1.1.0
